[tests-only] Only delete flock if file is no longer locked

+++ pkg/storage/utils/filelocks/filelocks.go
@@ -140,7 +140,9 @@ func ReleaseLock(lock *flock.Flock) error {
 
 	err = lock.Unlock()
 	if err == nil {
-		err = os.Remove(n)
+		if !lock.Locked() {
+			err = os.Remove(n)
+		}
 	}
 	releaseMutexedFlock(n)
